Split tests into two sections and remove redundant tests
